Cleanups for pedantic compiler warnings

+++ src/config.h
@@ -1,24 +1,24 @@
 #ifndef CONFIG_H
 #define CONFIG_H
 
-#define LCD1_GPIO_DC 6
-#define LCD1_GPIO_CS 7
-#define LCD2_GPIO_DC 8
-#define LCD2_GPIO_CS 9
-#define LCD3_GPIO_DC 2
-#define LCD3_GPIO_CS 3
-#define LCD4_GPIO_DC 4
-#define LCD4_GPIO_CS 5
-#define LCD5_GPIO_DC 14
-#define LCD5_GPIO_CS 15
-#define LCD6_GPIO_DC 16
-#define LCD6_GPIO_CS 17
-#define LCD7_GPIO_DC 18
-#define LCD7_GPIO_CS 19
-#define LCD_GPIO_RST 12
-#define LCD_GPIO_BL 13
-#define LCD_GPIO_CLK 10
-#define LCD_GPIO_MOSI 11
+#define LCD1_GPIO_DC (uint16_t)6
+#define LCD1_GPIO_CS (uint16_t)7
+#define LCD2_GPIO_DC (uint16_t)8
+#define LCD2_GPIO_CS (uint16_t)9
+#define LCD3_GPIO_DC (uint16_t)2
+#define LCD3_GPIO_CS (uint16_t)3
+#define LCD4_GPIO_DC (uint16_t)4
+#define LCD4_GPIO_CS (uint16_t)5
+#define LCD5_GPIO_DC (uint16_t)14
+#define LCD5_GPIO_CS (uint16_t)15
+#define LCD6_GPIO_DC (uint16_t)16
+#define LCD6_GPIO_CS (uint16_t)17
+#define LCD7_GPIO_DC (uint16_t)18
+#define LCD7_GPIO_CS (uint16_t)19
+#define LCD_GPIO_RST (uint16_t)12
+#define LCD_GPIO_BL (uint16_t)13
+#define LCD_GPIO_CLK (uint16_t)10
+#define LCD_GPIO_MOSI (uint16_t)11
 
 #define WATCHDOG_TIMEOUT_MS (20 * 1000)     // Watchdog timeout in milliseconds (needs to be longer than timeouts)
 #define WIFI_CONNECT_TIMEOUT_MS (10 * 1000) // Time to allow Wi-Fi driver to connect in milliseconds
